+++ docs/app/composables/useTheme.ts
@@ -103,11 +103,11 @@ export function useTheme() {
   const link = computed(() => [])
 
   const style = [
-    { innerHTML: radiusStyle, id: 'b24-ui-radius', tagPriority: -2 },
-    { innerHTML: blackAsPrimaryStyle, id: 'b24-ui-black-as-primary', tagPriority: -2 },
-    { innerHTML: fontStyle, id: 'b24-ui-font', tagPriority: -2 },
-    { innerHTML: customColorsStyle, id: 'chat-custom-colors', tagPriority: -2 },
-    { innerHTML: cssVariablesStyle, id: 'chat-css-variables', tagPriority: -2 }
+    { innerHTML: radiusStyle, id: 'b24-ui-radius', tagPriority: 'critical' as const },
+    { innerHTML: blackAsPrimaryStyle, id: 'b24-ui-black-as-primary', tagPriority: 'critical' as const },
+    { innerHTML: fontStyle, id: 'b24-ui-font', tagPriority: 'critical' as const },
+    { innerHTML: customColorsStyle, id: 'chat-custom-colors', tagPriority: 'critical' as const },
+    { innerHTML: cssVariablesStyle, id: 'chat-css-variables', tagPriority: 'critical' as const }
   ]
 
   const hasCSSChanges = computed(() => false)